backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
apscheduler
openpyxl
bcrypt
pdfplumber
python-docx
lxml
anthropic
